# Backend/openapi.yaml
openapi: 3.1.0
info:
  title: SCUD Backend API
  version: "1.0.0"
  description: |
    HTTP API системы управления доступом (SCUD).

    Контракт описывает три класса участников:

    - **Мобильное приложение пользователя** (Android) — авторизуется
      session_token'ом, выпускает ключи доступа, скачивает Bloom-фильтры для
      доставки на ридер, отправляет отчёты от ридера на сервер.
    - **Администратор / интеграция** — авторизуется API-ключом
      (`X-Api-Key`), управляет пользователями, группами, ридерами, permit'ами,
      выпущенными ключами, наблюдает доставку фильтров.
    - **Public** — health-check и пара unauthenticated auth-эндпоинтов
      (login / refresh).

    Ключевые понятия:

    - **Reader** — физический контроллер двери (ESP32). Имеет 16-байтный
      `reader_id` и Ed25519 публичный ключ.
    - **Reader group** — группа ридеров; пользователь видит только ридеры
      из доступной ему `user_group → reader_group` пары.
    - **Permit** — разрешение «пользователь P может пользоваться ридером R
      в окне [valid_from, valid_until], выпускать до n_parallel ключей с
      TTL ≤ max_token_ttl_seconds».
    - **IssuedKey** — выпущенный пользователем ключ (CARD на терминологию
      ридера). Подписан сервером, содержит `permit_id`, `device_id`, окно
      `[issued_at, expires_at]`. Отозванные ключи попадают в Bloom-фильтр
      ридера.
    - **TimeGrant** — подписанный сервером talisman, разрешающий телефону
...
      `"revoked"`).
servers:
...
security:
  - SessionToken: []
  - AdminApiKey: []
paths:
...
  /api/v1/app/auth/login:
    post:
      tags: [app-auth]
      summary: Логин по логину/паролю, получить session_token
      description: |
        Аутентифицирует пользователя по `login`/`password`. На успех создаёт
        запись Session (`session_token` живёт ~1 час, `refresh_token` —
        дольше) и возвращает оба токена + базовые данные пользователя.

        Если в запросе передан `device_info` — он пишется в Session как
        контекст устройства; сама регистрация ключа устройства (Ed25519
        pubkey) делается отдельно через `/auth/register-device`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/LoginRequest"}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AuthTokens"}
        "401":
          description: invalid_credentials
          content: {application/json: {schema: {$ref: "#/components/schemas/ErrorDetail"}}}
        "403":
          description: user_inactive
          content: {application/json: {schema: {$ref: "#/components/schemas/ErrorDetail"}}}
...
  /api/v1/app/auth/logout:
    post:
      tags: [app-auth]
      summary: Завершить текущую сессию
      description: |
        Помечает текущую Session как revoked. Refresh_token также становится
        невалидным.
      security:
        - SessionToken: []
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/Ok"}}}
        "401": {$ref: "#/components/responses/Unauthorized"}

  # ===================================================================
  # APP / READERS
  # ===================================================================
  /api/v1/app/readers:
    get:
      tags: [app-readers]
      summary: Список ридеров, доступных пользователю
      description: |
        Возвращает ридеры из reader_group'ов, привязанных к user_group'е
        текущего пользователя. Для каждого ридера приводит публичный ключ —
        приложение использует его при верификации подписей PUSH_INFO/FDI/BLK.
      security:
        - SessionToken: []
      parameters:
        - {name: group_id, in: query, required: false, schema: {type: string, format: uuid}, description: "Опциональный фильтр по reader_group."}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/ReaderPublic"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: "not_accessible (group вне доступной user_group)"}
  /api/v1/app/readers/{reader_id_hex}:
    get:
      tags: [app-readers]
      summary: Один ридер по hex ID
      description: Сводка по одному ридеру (если он доступен пользователю).
      security:
        - SessionToken: []
      parameters:
        - {name: reader_id_hex, in: path, required: true, schema: {$ref: "#/components/schemas/ReaderIdHex"}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ReaderPublic"}
        "400": {description: "invalid reader_id hex"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: "not_accessible"}
        "404": {description: "reader_not_found"}

  # ===================================================================
  # APP / PERMITS
  # ===================================================================
...
  /api/v1/admin/users:
    post:
      tags: [admin-users]
      summary: Создать пользователя
      description: |
        Заводит запись User (логин/пароль/display_name/user_group). Пароль
        хешируется (argon2/bcrypt). На дубликат логина — **409 login_taken**.
      security:
        - AdminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateUserRequest"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id: {type: integer}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "409": {description: "login_taken"}
    get:
      tags: [admin-users]
      summary: Список пользователей
      description: Постраничный список всех пользователей (cursor-offset).
      security:
        - AdminApiKey: []
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/AdminUser"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /api/v1/admin/users/{user_id}:
    get:
      tags: [admin-users]
      summary: Подробности пользователя
      security:
        - AdminApiKey: []
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/AdminUser"}}}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "user_not_found"}
    patch:
      tags: [admin-users]
      summary: Изменить пользователя
      description: |
        Меняет `display_name`, `user_group_id` или `is_active`. При
        `is_active=false` все активные сессии пользователя отзываются.
      security:
        - AdminApiKey: []
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PatchUserRequest"}
      responses:
        "200": {$ref: "#/components/responses/Ok"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "user_not_found"}
  /api/v1/admin/users/{user_id}/reset-password:
    post:
      tags: [admin-users]
      summary: Сбросить пароль
      description: Заменяет пароль пользователя; refresh-токены не инвалидируются автоматически.
      security:
        - AdminApiKey: []
      parameters:
        - {name: user_id, in: path, required: true, schema: {type: integer}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_password]
              properties:
                new_password: {type: string}
      responses:
        "200": {$ref: "#/components/responses/Ok"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "user_not_found"}

  # ===================================================================
  # ADMIN / READERS
  # ===================================================================
...
  /api/v1/admin/readers:
    get:
      tags: [admin-readers]
      summary: Список ридеров
      description: |
        Постраничный список ридеров с базовой телеметрией: время последнего
        контакта, последняя применённая версия фильтра. Может фильтроваться
        по `group_id`.
      security:
        - AdminApiKey: []
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
        - {name: group_id, in: query, required: false, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/AdminReader"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /api/v1/admin/readers/{reader_id_hex}:
    get:
      tags: [admin-readers]
      summary: Подробности ридера
      security:
        - AdminApiKey: []
      parameters:
        - {name: reader_id_hex, in: path, required: true, schema: {$ref: "#/components/schemas/ReaderIdHex"}}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/AdminReaderDetail"}}}
        "400": {description: "invalid hex"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "reader_not_found"}
    patch:
      tags: [admin-readers]
      summary: Изменить ридер
      description: Меняет display_name, description, reader_group_id или is_active.
      security:
        - AdminApiKey: []
      parameters:
        - {name: reader_id_hex, in: path, required: true, schema: {$ref: "#/components/schemas/ReaderIdHex"}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PatchReaderRequest"}
      responses:
        "200": {$ref: "#/components/responses/Ok"}
        "400": {description: "invalid hex"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "reader_not_found"}

  # ===================================================================
  # ADMIN / READER GROUPS
  # ===================================================================
  /api/v1/admin/reader-groups:
    post:
      tags: [admin-reader-groups]
      summary: Создать группу ридеров
      security:
        - AdminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: {type: string}
                description: {type: string, nullable: true}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  group_id: {type: string, format: uuid}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
    get:
      tags: [admin-reader-groups]
      summary: Список групп ридеров
      security:
        - AdminApiKey: []
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/ReaderGroup"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /api/v1/admin/reader-groups/{group_id}:
    patch:
      tags: [admin-reader-groups]
      summary: Изменить группу
      security:
        - AdminApiKey: []
      parameters:
        - {name: group_id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: {type: string}
                description: {type: string, nullable: true}
      responses:
        "200": {$ref: "#/components/responses/Ok"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "group_not_found"}
    delete:
      tags: [admin-reader-groups]
      summary: Удалить группу
      description: |
        Удаляет группу. Если в группе есть ридеры — **409 group_has_readers**.
      security:
        - AdminApiKey: []
      parameters:
        - {name: group_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200": {$ref: "#/components/responses/Ok"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "group_not_found"}
        "409": {description: "group_has_readers"}

  # ===================================================================
  # ADMIN / PERMITS
  # ===================================================================
...
components:
  securitySchemes:
    SessionToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        `Authorization: Bearer <session_token>` — выдаётся `/auth/login` или
        `/auth/refresh`. Хранится на сервере в таблице sessions, валидируется
        каждый запрос (revoked_at IS NULL, expires_at > now).
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        `X-Api-Key: <plaintext>` — выдаётся `/admin/api-keys` один раз при
        создании. На сервере хранится только SHA256-хеш + `key_prefix`.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: {type: integer, minimum: 0, default: 0}
      description: Offset для пагинации.
    Limit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Ok"}
    Unauthorized:
      description: missing_token / invalid_token / user_inactive
    AdminUnauthorized:
      description: missing_api_key / invalid_api_key / api_key_expired
  schemas:
    Ok:
      type: object
      properties:
        ok: {type: boolean, example: true}
    ErrorDetail:
      type: object
      description: |
        FastAPI стандартный формат — `{"detail": ...}`. detail может быть
        строкой (`"not_owner"`) или объектом (`{"error": "ttl_too_long",
        "max": 86400}`).
      properties:
        detail:
          oneOf:
            - {type: string}
            - {type: object}
    ReaderIdHex:
      type: string
      pattern: "^[0-9a-fA-F]{32}$"
      description: 16-байтный reader_id в hex (32 символа).
      example: "a1b2c3d4e5f60718293a4b5c6d7e8f90"

    # ----- AUTH -----
    LoginRequest:
      type: object
      required: [login, password]
      properties:
        login: {type: string}
        password: {type: string}
        device_info:
          type: object
          properties:
            model: {type: string}
            os: {type: string}
            app_version: {type: string}
...
    AuthTokens:
      type: object
      properties:
        session_token: {type: string}
        refresh_token: {type: string}
        user_id: {type: integer}
        user_group_id: {type: string, format: uuid}
        display_name: {type: string}

    # ----- READERS -----
    ReaderPublic:
      type: object
      properties:
        reader_id: {type: string, description: "hex"}
        display_name: {type: string}
        description: {type: string, nullable: true}
        reader_pubkey: {type: string, description: "base64, 32 bytes"}
        group_id: {type: string, format: uuid}
    AdminReader:
      type: object
      properties:
        reader_id: {type: string, description: "hex"}
        display_name: {type: string}
        description: {type: string, nullable: true}
        reader_group_id: {type: string, format: uuid}
        is_active: {type: boolean}
        enrolled_at: {type: string, format: date-time}
        last_contact_at: {type: string, format: date-time, nullable: true}
        last_applied_filter_version: {type: integer}
    AdminReaderDetail:
      allOf:
        - {$ref: "#/components/schemas/AdminReader"}
        - type: object
          properties:
            last_known_time: {type: string, format: date-time, nullable: true}
            open_delivery_tasks_count: {type: integer}
    PatchReaderRequest:
      type: object
      properties:
        display_name: {type: string}
        description: {type: string, nullable: true}
        reader_group_id: {type: string, format: uuid}
        is_active: {type: boolean}
...
    ReaderGroup:
      type: object
      properties:
        group_id: {type: string, format: uuid}
        name: {type: string}
        description: {type: string, nullable: true}
        created_at: {type: string, format: date-time}

    # ----- USERS (ADMIN) -----
    CreateUserRequest:
      type: object
      required: [login, password, display_name, user_group_id]
      properties:
        login: {type: string}
        password: {type: string}
        display_name: {type: string}
        user_group_id: {type: string, format: uuid}
    AdminUser:
      type: object
      properties:
        user_id: {type: integer}
        login: {type: string}
        display_name: {type: string}
        user_group_id: {type: string, format: uuid}
        is_active: {type: boolean}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    PatchUserRequest:
      type: object
      properties:
        display_name: {type: string}
        user_group_id: {type: string, format: uuid}
        is_active: {type: boolean}

    # ----- PERMITS (APP) -----
...